Bugfixing

+++ src/project4/string_util.h
@@ -11,7 +11,7 @@ public:
     if (x.length() != y.length()) {
       return false;
     }
-    for (int i = 0; i < x.length(); i++) {
+    for (size_t i = 0; i < x.length(); i++) {
       if (tolower(x[i]) != tolower(y[i])) {
         return false;
       }
